Start of splitting up outbound into separate files
